Moved external driver loading out of __init__ and into drivers.  Had to shuffle some import statements and move the BaseDriver/BaseCompressionDriver classes to gpsdio.base.  Closes #85.
